+++ scripts/02_train.py
@@ -38,7 +38,7 @@ def main(args):
     user_arg_dict["label_column"] = user_arg_dict.pop("label")
     user_arg_dict["timm_model_name"] = user_arg_dict.pop("model")
     user_arg_dict["class_map_name"] = user_arg_dict.pop("class_map")
-    return src.LightningModelWrapper(src.LightningModelArguments(**user_arg_dict)).train_model()
+    return src.TaxonomistModel(src.TaxonomistModelArguments(**user_arg_dict)).train_model()
 
 if __name__ == "__main__":
     trainer = main(sys.argv[1:]) # returns pytorch lightning trainer that has been trained.
