+++ src/taxonomist/__init__.py
@@ -1,6 +1,6 @@
 from .user_datasets import preprocess_dataset
 from .input_config import add_dataset_args, add_dataloader_args, add_model_args, add_train_args, add_program_args
 from .utils import load_class_map, get_pretrained_model_details, write_model_details_to_file, read_model_details_from_file
-from .lightning_data_wrapper import Dataset, LitDataModule, choose_aug
-from .lightning_model_wrapper import Model, LitModule, FeatureExtractionModule
-from .model_pipeline import LightningModelWrapper, LightningModelArguments
+from .data import Dataset, LitDataModule, choose_aug
+from .model import Model, LitModule, FeatureExtractionModule
+from .taxonomist_model import TaxonomistModel, TaxononomistModelArguments
